fix after master merge and review
