+++ src/main/java/faang/school/postservice/dto/post/PostDto.java
@@ -1,8 +1,7 @@
 package faang.school.postservice.dto.post;
 
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
