+++ src/main/java/faang/school/postservice/mapper/post/PostMapper.java
@@ -5,8 +5,11 @@ import faang.school.postservice.model.Post;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
 
+import java.util.List;
+
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface PostMapper {
-
-    PostDto toDto(Post entity);
+    PostDto toDto(Post post);
+    Post toEntity(PostDto post);
+    List<PostDto> toDtoList(List<Post> posts);
 }
