+++ src/main/java/faang/school/postservice/publisher/AbstractEventPublisher.java
@@ -5,7 +5,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.stereotype.Component;
 
 @Slf4j
@@ -24,9 +23,9 @@ public abstract class AbstractEventPublisher<T> implements MessagePublisher<T> {
             log.error("Error while creating event message", e);
             throw new RuntimeException(e);
         }
-        redisTemplate.convertAndSend(getChannelTopic().toString(), message);
+        redisTemplate.convertAndSend(getChannelTopic(), message);
         log.info("Published event {}", event);
     }
 
-    protected abstract ChannelTopic getChannelTopic();
+    protected abstract String getChannelTopic();
 }
